Fixing errors with circular dependencies
