+++ requirements_dev.txt
@@ -1,4 +1,4 @@
 -r requirements.txt
-pytest==7.1.2
-pytest-cov==3.0.9
-coverage==6.4
+pytest==8.3.3
+pytest-cov==5
+coverage==7.6.4
