+++ tests/expressions/test_sum.py
@@ -1,7 +1,7 @@
 import unittest
 
 from expressions.number import Number
-from expressions.sum import Sum
+from expressions.operations import Sum
 from expressions.variable import Variable
